renamed model-post and route-post, removed generateauthtoken function from model-post.
